+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## jannovar-core
 
+* refactorization in `AnnotationBuilderDispatcher` class
 * more refactorization, introducing `VariantAnnotation` class
 * moving around some functions in the `GenomeVariant` class hierarchy
 * renaming back `SmallVariantDescription` to `SmallVariantDescription`